# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Sales Partner API
  version: 1.0.0
  description: API for managing sales partners and partner contacts

servers:
  - url: http://localhost:8080/api

tags:
  - name: partner
    description: Operations related to sales partners
  - name: contact
    description: Operations related to partner contacts

paths:

  /partners:
    post:
      tags:
        - partner
      summary: Create partner
      operationId: createPartner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartnerRequest'
            example:
              name: Acme s.r.o.
              identifiers:
                - value: CZ12345678
                  type: ICO
                - value: CZ12345678
                  type: VAT
              addresses:
                - country: CZ
                  city: Prague
                  zipCode: "11000"
                  street: Wenceslas Square
                  streetNumber: "1"
      responses:
        '201':
          description: Partner created
          headers:
            Location:
              description: URL of the created partner
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerDetailResponse'
              example:
                id: "11111111-1111-1111-1111-111111111111"
                name: Acme s.r.o.
                partnerStatus: NEW
                identifiers:
                  - value: CZ12345678
                    type: ICO
                  - value: CZ12345678
                    type: VAT
                addresses:
                  - country: CZ
                    city: Prague
                    zipCode: "11000"
                    street: Wenceslas Square
                    streetNumber: "1"
                contacts: []
        '400':
          description: Invalid request or business validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner must contain at least one identifier.
                code: BUSINESS_VALIDATION_ERROR
        '409':
          description: Partner already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner with ICO CZ12345678 already exists.
                code: DUPLICATE_RESOURCE

    get:
      tags:
        - partner
      summary: Get all partners
      operationId: getPartners
      responses:
        '200':
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartnerListItem'
              example:
                - id: "11111111-1111-1111-1111-111111111111"
                  name: Acme s.r.o.
                  partnerStatus: NEW
                - id: "33333333-3333-3333-3333-333333333333"
                  name: Globex a.s.
                  partnerStatus: GOLD
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Unexpected server error.
                code: INTERNAL_SERVER_ERROR

  /partners/{partnerId}:
    get:
      tags:
        - partner
      summary: Get partner detail
      operationId: getPartnerDetail
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Partner detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerDetailResponse'
              example:
                id: "11111111-1111-1111-1111-111111111111"
                name: Acme s.r.o.
                partnerStatus: GOLD
                identifiers:
                  - value: CZ12345678
                    type: ICO
                  - value: CZ12345678
                    type: VAT
                addresses:
                  - country: CZ
                    city: Prague
                    zipCode: "11000"
                    street: Wenceslas Square
                    streetNumber: "1"
                contacts:
                  - id: "22222222-2222-2222-2222-222222222222"
                    firstName: Jan
                    lastName: Novak
                    position: SALES
                    primary: true
                    countryCallingCode: "+420"
                    phoneNumber: "777888999"
                    email: jan.novak@acme.example
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner with id 11111111-1111-1111-1111-111111111111 was not found.
                code: NOT_FOUND

    put:
      tags:
        - partner
      summary: Update partner
      operationId: updatePartner
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePartnerRequest'
            example:
              name: Acme Enterprise s.r.o.
              partnerStatus: GOLD
              identifiers:
                - value: CZ12345678
                  type: ICO
                - value: CZ12345678
                  type: VAT
              addresses:
                - country: CZ
                  city: Brno
                  zipCode: "60200"
                  street: Ceska
                  streetNumber: "15"
      responses:
        '200':
          description: Partner updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerDetailResponse'
              example:
                id: "11111111-1111-1111-1111-111111111111"
                name: Acme Enterprise s.r.o.
                partnerStatus: GOLD
                identifiers:
                  - value: CZ12345678
                    type: ICO
                  - value: CZ12345678
                    type: VAT
                addresses:
                  - country: CZ
                    city: Brno
                    zipCode: "60200"
                    street: Ceska
                    streetNumber: "15"
                contacts:
                  - id: "22222222-2222-2222-2222-222222222222"
                    firstName: Jan
                    lastName: Novak
                    position: ACCOUNT_MANAGER
                    primary: true
                    countryCallingCode: "+420"
                    phoneNumber: "777888999"
                    email: jan.novak@acme.example
        '400':
          description: Invalid request or business validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner name must not be blank.
                code: VALIDATION_ERROR
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner with id 11111111-1111-1111-1111-111111111111 was not found.
                code: NOT_FOUND
        '409':
          description: Partner conflicts with an existing resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner with ICO CZ12345678 already exists.
                code: DUPLICATE_RESOURCE

  /partners/{partnerId}/contacts:
    post:
      tags:
        - contact
      summary: Add contact to partner
      operationId: addContact
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
            example:
              firstName: Petra
              lastName: Dvorakova
              position: SUPPORT
              isPrimary: false
              countryCallingCode: "+420"
              phoneNumber: "666555444"
              email: petra.dvorakova@acme.example
      responses:
        '201':
          description: Contact created
          headers:
            Location:
              description: URL of the created contact
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
              example:
                id: "44444444-4444-4444-4444-444444444444"
                firstName: Petra
                lastName: Dvorakova
                position: SUPPORT
                primary: false
                countryCallingCode: "+420"
                phoneNumber: "666555444"
                email: petra.dvorakova@acme.example
        '400':
          description: Invalid request or business validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner can have only one primary contact.
                code: BUSINESS_VALIDATION_ERROR
        '404':
          description: Partner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Partner with id 11111111-1111-1111-1111-111111111111 was not found.
                code: NOT_FOUND
        '409':
          description: Contact already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Contact already exists for partner 11111111-1111-1111-1111-111111111111.
                code: DUPLICATE_RESOURCE

  /partners/{partnerId}/contacts/{contactId}:
    put:
      tags:
        - contact
      summary: Update contact
      operationId: updateContact
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactRequest'
            example:
              firstName: Petr
              lastName: Svoboda
              position: ACCOUNT_MANAGER
              isPrimary: false
              countryCallingCode: "+420"
              phoneNumber: "111222333"
              email: petr.svoboda@acme.example
      responses:
        '200':
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
              example:
                id: "22222222-2222-2222-2222-222222222222"
                firstName: Petr
                lastName: Svoboda
                position: ACCOUNT_MANAGER
                primary: false
                countryCallingCode: "+420"
                phoneNumber: "111222333"
                email: petr.svoboda@acme.example
        '400':
          description: Invalid request or business validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Contact firstName must not be blank.
                code: VALIDATION_ERROR
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Contact with id 22222222-2222-2222-2222-222222222222 was not found for partner 11111111-1111-1111-1111-111111111111.
                code: NOT_FOUND
        '409':
          description: Contact conflicts with an existing resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Contact already exists for partner 11111111-1111-1111-1111-111111111111.
                code: DUPLICATE_RESOURCE

    delete:
      tags:
        - contact
      summary: Remove or deactivate contact
      operationId: deleteContact
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Contact removed or deactivated
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Contact with id 22222222-2222-2222-2222-222222222222 was not found for partner 11111111-1111-1111-1111-111111111111.
                code: NOT_FOUND

components:

  schemas:

    CreatePartnerRequest:
      type: object
      required: [name]
      example:
        name: Acme s.r.o.
        identifiers:
          - value: CZ12345678
            type: ICO
        addresses:
          - country: CZ
            city: Prague
            zipCode: "11000"
            street: Wenceslas Square
            streetNumber: "1"
      properties:
        name:
          type: string
          example: Acme s.r.o.
          minLength: 2
          maxLength: 255
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/Identifier'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'

    UpdatePartnerRequest:
      allOf:
        - $ref: '#/components/schemas/CreatePartnerRequest'
        - type: object
          properties:
            partnerStatus:
              $ref: '#/components/schemas/PartnerStatus'

    PartnerListItem:
      type: object
      required: [ id, name, partnerStatus ]
      example:
        id: "11111111-1111-1111-1111-111111111111"
        name: Acme s.r.o.
        partnerStatus: NEW
      properties:
        id:
          type: string
          format: uuid
          example: "11111111-1111-1111-1111-111111111111"
        name:
          type: string
          example: Acme s.r.o.
          minLength: 1
          maxLength: 255
        partnerStatus:
          $ref: '#/components/schemas/PartnerStatus'

    PartnerDetailResponse:
      type: object
      required: [ id, name, partnerStatus ]
      example:
        id: "11111111-1111-1111-1111-111111111111"
        name: Acme s.r.o.
        partnerStatus: NEW
        identifiers:
          - value: CZ12345678
            type: ICO
        addresses:
          - country: CZ
            city: Prague
            zipCode: "11000"
            street: Wenceslas Square
            streetNumber: "1"
        contacts:
          - id: "22222222-2222-2222-2222-222222222222"
            firstName: Jan
            lastName: Novak
            position: SALES
            primary: true
            countryCallingCode: "+420"
            phoneNumber: "777888999"
            email: jan.novak@acme.example
          - id: "44444444-4444-4444-4444-444444444444"
            firstName: Petra
            lastName: Dvorakova
            position: SUPPORT
            primary: false
            countryCallingCode: "+420"
            phoneNumber: "666555444"
            email: petra.dvorakova@acme.example
      properties:
        id:
          type: string
          format: uuid
          example: "11111111-1111-1111-1111-111111111111"
        name:
          type: string
          example: Acme s.r.o.
          minLength: 2
          maxLength: 255
        partnerStatus:
          $ref: '#/components/schemas/PartnerStatus'
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/Identifier'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ContactResponse'

    CreateContactRequest:
      type: object
      required: [firstName, lastName]
      example:
        firstName: Petra
        lastName: Dvorakova
        position: SUPPORT
        isPrimary: false
        countryCallingCode: "+420"
        phoneNumber: "666555444"
        email: petra.dvorakova@acme.example
      properties:
        firstName:
          type: string
          example: Jan
          minLength: 2
          maxLength: 100
        lastName:
          type: string
          example: Novak
          minLength: 2
          maxLength: 100
        position:
          $ref: '#/components/schemas/ContactPosition'
        isPrimary:
          type: boolean
          example: true
        countryCallingCode:
          type: string
          example: "+420"
          minLength: 4
          maxLength: 4
          pattern: '^\+[0-9]{3}$'
          default: "+420"
        phoneNumber:
          type: string
          example: "777888999"
          minLength: 9
          maxLength: 9
          pattern: '^[0-9]{9}$'
        email:
          type: string
          example: jan.novak@acme.example
          format: email
          maxLength: 254
          pattern: '^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'

    UpdateContactRequest:
      $ref: '#/components/schemas/CreateContactRequest'

    ContactResponse:
      type: object
      required: [id, firstName, lastName]
      example:
        id: "44444444-4444-4444-4444-444444444444"
        firstName: Petra
        lastName: Dvorakova
        position: SUPPORT
        primary: false
        countryCallingCode: "+420"
        phoneNumber: "666555444"
        email: petra.dvorakova@acme.example
      properties:
        id:
          type: string
          format: uuid
          example: "22222222-2222-2222-2222-222222222222"
        firstName:
          type: string
          example: Jan
          minLength: 2
          maxLength: 100
        lastName:
          type: string
          example: Novak
          minLength: 2
          maxLength: 100
        position:
          $ref: '#/components/schemas/ContactPosition'
        primary:
          type: boolean
          example: true
        countryCallingCode:
          type: string
          example: "+420"
          minLength: 4
          maxLength: 4
          pattern: '^\+[0-9]{3}$'
          default: "+420"
        phoneNumber:
          type: string
          example: "777888999"
          minLength: 9
          maxLength: 9
          pattern: '^[0-9]{9}$'
        email:
          type: string
          example: jan.novak@acme.example
          format: email
          maxLength: 254
          pattern: '^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'

    Address:
      type: object
      required: [country, city]
      example:
        country: CZ
        city: Prague
        zipCode: "11000"
        street: Wenceslas Square
        streetNumber: "1"
      properties:
        country:
          type: string
          example: CZ
          minLength: 2
          maxLength: 2
          pattern: '^[A-Z]{2}$'
        city:
          type: string
          example: Prague
          minLength: 1
          maxLength: 100
        zipCode:
          type: string
          example: "11000"
          minLength: 5
          maxLength: 10
          pattern: '^[0-9A-Za-z -]+$'
        street:
          type: string
          example: Wenceslas Square
          minLength: 1
          maxLength: 150
        streetNumber:
          type: string
          example: "1"
          minLength: 1
          maxLength: 20

    Identifier:
      type: object
      required: [value, type]
      example:
        value: CZ12345678
        type: ICO
      properties:
        value:
          type: string
          example: CZ12345678
          minLength: 1
          maxLength: 50
        type:
          $ref: '#/components/schemas/IdentifierType'

    IdentifierType:
      type: string
      enum: [ICO, VAT, OTHER]

    ContactPosition:
      type: string
      enum: [EXECUTIVE, SALES, ACCOUNT_MANAGER, TECHNICAL, FINANCE, SUPPORT, OTHER]

    PartnerStatus:
      type: string
      enum: [NEW, STANDARD, GOLD, INACTIVE]

    ErrorResponse:
      type: object
      required: [ message ]
      example:
        message: Partner with id 999 was not found.
        code: PARTNER_NOT_FOUND
      properties:
        message:
          type: string
          example: Partner with id 999 was not found.
          minLength: 1
          maxLength: 500
        code:
          type: string
          example: PARTNER_NOT_FOUND
          minLength: 1
          maxLength: 100